+++ CHANGES.rst
@@ -12,6 +12,7 @@ in progress
 
 - Prepare Saraswati to run as system service. There is now a corresponding
   command ``sudo saraswati setup --systemd`` which aids this.
+-  Emit warning when no audio pipelines are defined
 
 
 2021-06-21 0.3.2